如果採用，此提交將嘗試使用條件判斷式，待下次新系統時開發測試使用
修改項目:

將使用Docker官方提供腳本安裝

模組:

議題

+++ Docker.sh
@@ -1,18 +1,29 @@
 # 安裝docker
-sudo apt-get update -y
-sudo apt-get upgrade -y
-sudo apt-get remove docker docker-engine docker.io containerd runc -y
-sudo apt-get install apt-transport-https ca-certificates curl gnupg-agent software-properties-common -y -f
-curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add -
-sudo apt-key fingerprint 0EBFCD88
-sudo add-apt-repository "deb [arch=amd64] https://download.docker.com/linux/ubuntu $(lsb_release -cs) stable"
-sudo apt-get update -y
-sudo apt-get install docker-ce docker-ce-cli containerd.io -y -f
-sudo apt-cache madison docker-ce -y
+if docker > /dev/null
+  then
+    sudo apt-get update -y
+    sudo apt-get upgrade -y
+    sudo apt-get remove docker docker-engine docker.io containerd runc -y
+    sudo apt-get install apt-transport-https ca-certificates curl gnupg-agent software-properties-common -y -f
+    curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add -
+    sudo apt-key fingerprint 0EBFCD88
+    sudo add-apt-repository "deb [arch=amd64] https://download.docker.com/linux/ubuntu $(lsb_release -cs) stable"
+    echo \
+      "deb [arch=amd64 signed-by=/usr/share/keyrings/docker-archive-keyring.gpg] https://download.docker.com/linux/ubuntu \
+      $(lsb_release -cs) stable" | sudo tee /etc/apt/sources.list.d/docker.list > /dev/null
+    sudo apt-get update -y
+    sudo apt-get install docker-ce docker-ce-cli containerd.io -y -f
+    sudo apt-cache madison docker-ce -y
+  else
+    curl -fsSL https://get.docker.com -o get-docker.sh
+    sudo sh get-docker.sh
+    sudo groupadd docker
+    sudo usermod -aG docker $USER
 
-sudo groupadd docker
-sudo usermod -aG docker $USER
+    sudo systemctl enable docker
+    sudo systemctl enable docker.service
+    sudo systemctl enable containerd.service
 
-sudo systemctl enable docker
+  fi
 
 docker run hello-world
